Add files via upload
The Eye Bank Management System is designed to streamline and automate the entire process of eye donation, stock management, and patient requests.                            By digitizing and automating the management of eye donation and transplantation, the Eye Bank Management System enhances the overall efficiency and transparency of the process. This solution not only helps reduce manual errors but also accelerates the process of matching donors to recipients, ensuring that patients receive the medical help they need in a timely manner. Through its intuitive interface, the system provides eye banks and hospitals with a reliable tool to improve the quality of corneal transplant services, ultimately helping restore vision to those in need.